+++ Source/ArticyEditor/Private/ArticyEditorModule.cpp
@@ -25,6 +25,7 @@
 #include "IDirectoryWatcher.h"
 #include "Customizations/ArticyPinFactory.h"
 #include "Customizations/AssetActions/AssetTypeActions_ArticyGV.h"
+#include "Customizations/AssetActions/AssetTypeActions_ArticyAlterativeGV.h"
 #include "Customizations/Details/ArticyGVCustomization.h"
 #include "Customizations/Details/ArticyPluginSettingsCustomization.h"
 #include "Customizations/Details/ArticyIdCustomization.h"
@@ -184,6 +185,7 @@ void FArticyEditorModule::RegisterAssetTypeActions()
 {
 	IAssetTools& AssetTools = FModuleManager::LoadModuleChecked<FAssetToolsModule>("AssetTools").Get();
 	AssetTools.RegisterAssetTypeActions(MakeShareable(new FAssetTypeActions_ArticyGV()));
+	AssetTools.RegisterAssetTypeActions(MakeShareable(new FAssetTypeActions_ArticyAlterativeGV()));
 }
 
 void FArticyEditorModule::RegisterPluginCommands()