+++ Source/ArticyEditor/Private/CodeGeneration/ExpressoScriptsGenerator.cpp
@@ -88,19 +88,27 @@ void GenerateExpressoScripts(CodeFileGenerator* header, const UArticyImportData*
 	auto gvTypeName = CodeGenerator::GetGlobalVarsClassname(Data);
 	for(const auto ns : Data->GetGlobalVars().Namespaces)
 		header->Variable("mutable TWeakObjectPtr<" + ns.CppTypename + ">", ns.Namespace, "nullptr");
+	header->Variable("mutable TWeakObjectPtr<" + gvTypeName + ">", "ActiveGlobals", "nullptr");
 
 	header->Line();
 	header->Method("void", "SetGV", "UArticyGlobalVariables* GV", [&]
 	{
 		header->Variable("auto", "gv", FString::Printf(TEXT("Cast<%s>(GV)"), *gvTypeName));
-		header->Line("if(ensure(gv))");
-		header->Block(true, [&]
-		{
-			header->Comment("Initialize all GV namespace references");
-			for(const auto ns : Data->GetGlobalVars().Namespaces)
-				header->Line(FString::Printf(TEXT("%s = gv->%s;"), *ns.Namespace, *ns.Namespace));
-		});
+		header->Comment("Initialize all GV namespace references (or nullify if we're setting to nullptr)");
+		for (const auto ns : Data->GetGlobalVars().Namespaces)
+			header->Line(FString::Printf(TEXT("%s = gv ? gv->%s : nullptr;"), *ns.Namespace, *ns.Namespace));
+
+		header->Comment("Store GVs");
+		header->Line("ActiveGlobals = gv;");
 	}, "", false, "", "const override");
+
+	header->Line();
+	header->Method("UArticyGlobalVariables*", "GetGV", "", [&]
+	{
+		header->Comment("Return active global variables as set by SetGV");
+		header->Line("if(ActiveGlobals.IsValid()) { return ActiveGlobals.Get(); }");
+		header->Line("return nullptr;");
+	}, "", false, "", "override");
 
 	header->Line();
 	header->Method("UClass*", "GetUserMethodsProviderInterface", "", [&]
@@ -109,23 +117,15 @@ void GenerateExpressoScripts(CodeFileGenerator* header, const UArticyImportData*
 	}, "", false, "", "override");
 
 	header->Line();
-	header->Method("UObject*", "GetUserMethodsProviderObject", "", [&]
-	{
-		header->Line("if(UserMethodsProvider)");
-		header->Line("	return UserMethodsProvider;");
-		header->Line("if(DefaultUserMethodsProvider)");
-		header->Line("	return DefaultUserMethodsProvider;");
-		header->Line("return nullptr;");
-	}, "", false, "", "const");
-
-	header->Line();
 	header->Line("public:", false, true, -1);
 
 	header->Line();
 	// disable "optimization cannot be applied due to function size" compile error. This error is caused by the huge constructor when all expresso
 	// scripts are added to the collection and this pragma disables the optimizations. 
+	header->Line("#if !((defined(PLATFORM_PS4) && PLATFORM_PS4) || (defined(PLATFORM_PS5) && PLATFORM_PS5))");
 	header->Line("#pragma warning(push)");
 	header->Line("#pragma warning(disable: 4883) //<disable \"optimization cannot be applied due to function size\" compile error.");
+	header->Line("#endif");
 	header->Method("", CodeGenerator::GetExpressoScriptsClassname(Data), "", [&]
 	{
 		const auto fragments = Data->GetScriptFragments();
@@ -164,7 +164,9 @@ void GenerateExpressoScripts(CodeFileGenerator* header, const UArticyImportData*
 			}
 		}
 	});
+	header->Line("#if !((defined(PLATFORM_PS4) && PLATFORM_PS4) || (defined(PLATFORM_PS5) && PLATFORM_PS5))");
 	header->Line("#pragma warning(pop)");
+	header->Line("#endif");
 }
 
 void ExpressoScriptsGenerator::GenerateCode(const UArticyImportData* Data)
@@ -175,6 +177,7 @@ void ExpressoScriptsGenerator::GenerateCode(const UArticyImportData* Data)
 
 	CodeFileGenerator(GetFilename(Data), true, [&](CodeFileGenerator* header)
 	{
+		header->Line("#include \"CoreUObject.h\"");
 		header->Line("#include \"ArticyRuntime/Public/ArticyExpressoScripts.h\"");
 		header->Line("#include \"" + CodeGenerator::GetGlobalVarsClassname(Data, true) + ".h\"");
 		header->Line("#include \"" + CodeGenerator::GetExpressoScriptsClassname(Data, true) + ".generated.h\"");