+++ Source/ArticyEditor/Private/Slate/AssetPicker/SArticyObjectAssetPicker.cpp
@@ -200,7 +200,11 @@ TSharedRef<SWidget> SArticyObjectAssetPicker::CreateClassPicker()
 	FClassViewerInitializationOptions ClassViewerConfig;
 	ClassViewerConfig.DisplayMode = EClassViewerDisplayMode::ListView;
 	ClassViewerConfig.bAllowViewOptions = true;
+#if ENGINE_MAJOR_VERSION >= 5
+	ClassViewerConfig.ClassFilters.Add(MakeShareable(new FArticyRefClassFilter(TopLevelClassRestriction.Get(), false)));
+#else
 	ClassViewerConfig.ClassFilter = MakeShareable(new FArticyRefClassFilter(TopLevelClassRestriction.Get(), false));
+#endif
 
 	return FModuleManager::LoadModuleChecked<FClassViewerModule>("ClassViewer").CreateClassViewer(ClassViewerConfig, FOnClassPicked::CreateRaw(this, &SArticyObjectAssetPicker::OnClassPicked_Func));
 }
@@ -228,7 +232,9 @@ TSharedRef<class ITableRow> SArticyObjectAssetPicker::MakeTileViewWidget(TWeakOb
 {
 	TSharedPtr< STableRow<TWeakObjectPtr<UArticyObject>>> TableRowWidget;
 	SAssignNew(TableRowWidget, STableRow<TWeakObjectPtr<UArticyObject>>, OwnerTable)
+#if ENGINE_MAJOR_VERSION == 4
 		.Style(FEditorStyle::Get(), "ContentBrowser.AssetListView.TableRow")
+#endif
 		.Cursor(true ? EMouseCursor::GrabHand : EMouseCursor::Default)
 		.Padding(3.f);