+++ Source/ArticyRuntime/Private/ArticyFlowPlayer.cpp
@@ -120,16 +120,19 @@ UArticyDatabase* UArticyFlowPlayer::GetDB() const
 
 UArticyGlobalVariables* UArticyFlowPlayer::GetGVs() const
 {
-	return OverrideGV ? OverrideGV : UArticyGlobalVariables::GetDefault(this);
+	// If we have an custom GV set, make sure we're using a runtime clone of it
+	if (OverrideGV) {
+		return UArticyGlobalVariables::GetRuntimeClone(this, OverrideGV);
+	}
+	
+	// Otherwise, use the default variable set
+	return UArticyGlobalVariables::GetDefault(this);
 }
 
 UObject* UArticyFlowPlayer::GetMethodsProvider() const
 {
 	auto expressoInstance = GetDB()->GetExpressoInstance();
 	auto provider = expressoInstance->GetUserMethodsProviderInterface();
-	
-	if (expressoInstance->UserMethodsProvider != nullptr && UserMethodsProvider == nullptr)//MM_CHANGE
-		UserMethodsProvider = expressoInstance->UserMethodsProvider;
 
 	if(ensure(provider))
 	{
@@ -164,7 +167,7 @@ UObject* UArticyFlowPlayer::GetMethodsProvider() const
 						}
 
 						//and finally we check for a default methods provider, which we can use as fallback
-						auto defaultUserMethodsProvider = expressoInstance->DefaultUserMethodsProvider;
+						auto defaultUserMethodsProvider = expressoInstance->GetDefaultUserMethodsProvider();
 						if(defaultUserMethodsProvider && ensure(defaultUserMethodsProvider->GetClass()->ImplementsInterface(provider)))
 							UserMethodsProvider = defaultUserMethodsProvider;
 					}