+++ Source/ArticyRuntime/Public/ArticyAlternativeGlobalVariables.h
@@ -1,0 +1,14 @@
+//  
+// Copyright (c) articy Software GmbH & Co. KG. All rights reserved.  
+//
+#pragma once
+
+#include "Engine/DataAsset.h"
+#include "ArticyAlternativeGlobalVariables.generated.h"
+
+UCLASS(BlueprintType)
+class ARTICYRUNTIME_API UArticyAlternativeGlobalVariables : public UDataAsset
+{
+    GENERATED_BODY()
+public:
+};