+++ Source/ArticyRuntime/Public/ArticyBaseObject.h
@@ -29,8 +29,6 @@ public:
 	/** For ArticyImporter internal use only: initializes the bWasDeserialized variable. */
 	void Initialize() { bWasDeserialized = true; }
 
-	UWorld* GetWorld() const override { return GEngine->GetWorldFromContextObjectChecked(GetOuter()); }
-
 	UArticyPrimitive* GetSubobject(FArticyId Id) const;
 
 protected: