+++ Changelog.md
@@ -1,4 +1,26 @@
-## Unreal Import 1.3.1 Changelog
+## Unreal Importer 1.4 Changelog
+
+- Unreal Engine 5 support
+- Breaking Changes:
+  - A change in the code generator will break existing projects. Search "Error C2451" in [the readme](README.md) for the quick fix to get you compiling again.
+- New Features:
+  - Rich text support using Unreal's Rich Text Block component, including hyperlinks ([#64](https://github.com/ArticySoftware/ArticyImporterForUnreal/pull/64)).
+  - Support for multiple, independent global variable sets ([#66](https://github.com/ArticySoftware/ArticyImporterForUnreal/pull/66)).
+- Changes:
+  - Import with enabled live coding is now allowed in UE5.
+  - Moved generated method `U<ProjectName>ExpressoScripts::GetUserMethodsProviderObject()` to `UArticyExpressoScripts::GetUserMethodsProviderObject()`.
+- Fixes:
+  - Fixed compilation issues with UE4.22
+  - Fixed compilation issues with Linux cross-compilation toolchain v19
+  - Fixed packaging compiler warning concerning `FADIHierarchy::RootObject` ([#60](https://github.com/ArticySoftware/ArticyImporterForUnreal/pull/60)).
+  - Fixed unsupported pragma compile error when building for PlayStation 4/5 ([#59](https://github.com/ArticySoftware/ArticyImporterForUnreal/pull/59) and [#67](https://github.com/ArticySoftware/ArticyImporterForUnreal/pull/67)).
+  - Fixed issue with Unicode characters in generated scripts.
+  - Fixed crash when changing levels while using a custom script methods ([#63](https://github.com/ArticySoftware/ArticyImporterForUnreal/pull/63)).
+  - Fixed issues with `GetWorld()` in `UArticyDatabase` and `UArticyBaseObject` ([#68](https://github.com/ArticySoftware/ArticyImporterForUnreal/issues/68)).
+- Documentation
+  - Added documentation for custom script methods and shadowing to the Readme ([#61](https://github.com/ArticySoftware/ArticyImporterForUnreal/pull/61)).
+
+## Unreal Importer 1.3.1 Changelog
 
 - General:
     - New warning dialogue if you try to run the importer and a hotload is required while Live Coding is enabled ([#53](https://github.com/ArticySoftware/ArticyImporterForUnreal/pull/53)).
@@ -13,14 +35,14 @@
     - Added proper dependencies so `ArticyEditorModule` can be included in another module without errors ([#57](https://github.com/ArticySoftware/ArticyImporterForUnreal/pull/57)).
 
 
-## Unreal Import 1.3.0 Changelog
+## Unreal Importer 1.3.0 Changelog
 - Unreal Engine 5 Early Access 2 Support ([#41](https://github.com/ArticySoftware/ArticyImporterForUnreal/pull/41)).
 
 - Fixes:
     - Fixed issue with plugin marking database and package assets for delete. Fixed in [#42](https://github.com/ArticySoftware/ArticyImporterForUnreal/pull/42). Resolves [#39](https://github.com/ArticySoftware/ArticyImporterForUnreal/issues/39).
     - Support for Unreal 4.27
 
-## Unreal Import 1.2.0 Changelog
+## Unreal Importer 1.2.0 Changelog
 - Unreal Engine 4.26 Support
 - Unreal Engine 4.20 and 4.21 no longer supported
 - New Features: